Quite a bit of edits here on this commit as I am learning how much I can control with flexboxes.
HTML: for the html I have added the page links and they're fully working. The coinciding CSS is there with the correct code.

CSS: For the CSS I added a lot of comments to the forsure code that needs to be there. Or at least for now seems I need for the profile I am looking for. And for the page links I added the needed flex coding.